recap-subworker/pyproject.toml
[project]
name = "recap-subworker"
version = "0.1.0"
description = "Alt recap evidence clustering subworker"
requires-python = "~=3.12"
readme = "README.md"
authors = [{ name = "Alt Engineering" }]
dependencies = [
  "fastapi>=0.115.12",
  "uvicorn[standard]>=0.32.0",
  "gunicorn>=22.0.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.3.0",
  "orjson==3.11.4",
  "prometheus-fastapi-instrumentator==7.1.0",
  "fastjsonschema==2.21.2",
  "sentence-transformers==5.1.2",
  "numpy>=2.1.0",
  "scikit-learn>=1.5.0",
  "umap-learn==0.5.9.post2",
  "hdbscan==0.8.40",
  "fast-hdbscan==0.2.1",
  "scipy>=1.13.0",
  "structlog>=24.1.0",
  "sqlalchemy[asyncio]>=2.0.34",
  "asyncpg>=0.29.0",
  "nltk>=3.8.1",
  "spacy>=3.7.0"
]

[project.optional-dependencies]
distill = ["sentence-transformers==5.1.2"]
onnx = [
  "onnxruntime>=1.19.0",
  "onnxruntime-tools>=1.7.0"
]

[project.scripts]
recap-subworker = "recap_subworker.__main__:main"

[tool.uv]
package = true
dev-dependencies = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "mypy>=1.11.0",
  "ruff>=0.6.0"
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]
markers = [
  "perf: performance tests"
]
